% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_shapes.R
\name{calculate_shape_area}
\alias{calculate_shape_area}
\title{Calculate shape area}
\usage{
calculate_shape_area(shape_type, params)
}
\arguments{
\item{shape_type}{must be one of: 'square', 'circle', 'rectangle'}

\item{params}{list of paramters (see details of process function)}
}
\value{
single numeric value representing shape area
}
\description{
Calculate shape area
}
