% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_shapes.R
\name{parse_shape_definition_line}
\alias{parse_shape_definition_line}
\title{Parse a single line of shape definitions string into a shape definition}
\usage{
parse_shape_definition_line(shape_definition_line)
}
\arguments{
\item{shape_definition_line}{a string that defines a single shape
(see details of process function)}
}
\value{
a list with elements shape_name, shape_type, and parameters:w
}
\description{
Parse a single line of shape definitions string into a shape definition
}
