% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_shapes.R
\name{process}
\alias{process}
\title{Calculate Area of Shapes}
\usage{
process(shape_definitions_string)
}
\arguments{
\item{shape_definitions_string}{a string where each line defines a shape like
`shape_name: shape_type (shape_params)`. (See Details)}
}
\value{
dataframe with columns: `name`, `type`, and `area` name and type
 coming from input definitions and area calculated based on params.
}
\description{
Calculate Area of Shapes
}
\details{
`shape_name` can be any alpha numeric string, but should not contain
whitespace or symbols `shape_type` must be one of square, circle, and
rectangle. `shape_params` must be comma separated `param_name=param_values`
Each shape has required shape params as follows: .

.   - `square`: `l` (length)
.   - `rectangle`: `l` (length), `w` (width) .
.   - `circle`: `r` (radius)
}
